// whatsup-client/src/services/models/ChatResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ChatResponse
 */
export interface ChatResponse {
    /**
     * 
     * @type {string}
     * @memberof ChatResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatResponse
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof ChatResponse
     */
    unreadCount?: number;
    /**
     * 
     * @type {string}
     * @memberof ChatResponse
     */
    lastMessage?: string;
    /**
     * 
     * @type {Date}
     * @memberof ChatResponse
     */
    lastMessageTime?: Date;
    /**
     * 
     * @type {string}
     * @memberof ChatResponse
     */
    senderId?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatResponse
     */
    receiverId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ChatResponse
     */
    receiverOnline?: boolean;
}

/**
 * Check if a given object implements the ChatResponse interface.
 */
export function instanceOfChatResponse(value: object): value is ChatResponse {
    return true;
}

export function ChatResponseFromJSON(json: any): ChatResponse {
    return ChatResponseFromJSONTyped(json, false);
}

export function ChatResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChatResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'unreadCount': json['unreadCount'] == null ? undefined : json['unreadCount'],
        'lastMessage': json['lastMessage'] == null ? undefined : json['lastMessage'],
        'lastMessageTime': json['lastMessageTime'] == null ? undefined : (new Date(json['lastMessageTime'])),
        'senderId': json['senderId'] == null ? undefined : json['senderId'],
        'receiverId': json['receiverId'] == null ? undefined : json['receiverId'],
        'receiverOnline': json['receiverOnline'] == null ? undefined : json['receiverOnline'],
    };
}

export function ChatResponseToJSON(json: any): ChatResponse {
    return ChatResponseToJSONTyped(json, false);
}

export function ChatResponseToJSONTyped(value?: ChatResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'unreadCount': value['unreadCount'],
        'lastMessage': value['lastMessage'],
        'lastMessageTime': value['lastMessageTime'] == null ? undefined : ((value['lastMessageTime']).toISOString()),
        'senderId': value['senderId'],
        'receiverId': value['receiverId'],
        'receiverOnline': value['receiverOnline'],
    };
}